+++ Assets/Code/Scripts/Weapon.cs
@@ -3,7 +3,7 @@
     // 모든 종류의 무기에 공통적으로 들어가는 요소들을 지정합니다.
     public class Weapon
     {
-        private int _dmg; // 공격력
+        private float _dmg; // 공격력
         private float _atkSpeed; // 공격 속도
         private float _atkRate; // 지속 피해 주기
         private float _durability; // 내구도
